Added bottom navigation view, moved all code to fragments, added manage fragment
