+++ app/src/main/java/com/tudelft/smartphonesensing/TestFragment.java
@@ -1,0 +1,20 @@
+package com.tudelft.smartphonesensing;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+public class TestFragment extends Fragment {
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+
+        return inflater.inflate(R.layout.test_fragment, container, false);
+    }
+}
