+++ app/src/main/res/layout/manage_fragment.xml
@@ -1,0 +1,53 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    >
+
+
+
+    <TextView
+        android:padding="20dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Enter Number of Cells:"
+        android:textSize="20dp">
+
+    </TextView>
+
+    <NumberPicker
+    android:layout_margin="20dp"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:id="@+id/numPicker"
+    />
+
+    <TextView
+    android:padding="20dp"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:text="Enter Wifi MACS (comma seperated):"
+    android:textSize="20dp"/>
+
+  <EditText
+      android:layout_width="match_parent"
+      android:layout_height="120dp"
+      android:layout_margin="20dp"
+      android:background="@color/cardview_light_background"
+      android:id="@+id/edit_tv"/>
+
+
+
+
+
+
+
+
+
+
+
+
+</LinearLayout>
