+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/navigation_manage"
+        android:enabled="true"
+        android:icon="@android:drawable/ic_menu_manage"
+        android:title="Manage"/>
+    <item
+        android:id="@+id/navigation_train"
+        android:enabled="true"
+        android:icon="@android:drawable/ic_input_add"
+        android:title="Collect Data"/>
+    <item
+        android:id="@+id/navigation_test"
+        android:enabled="true"
+        android:icon="@android:drawable/ic_menu_mylocation"
+        android:title="Predict Location"/>
+</menu>
